requirements.txt
beautifulsoup4==4.13.4
fastapi==0.116.1
gunicorn
googlesearch_python==1.3.0
indic_transliteration==2.3.72
joblib==1.5.1
lyricsgenius==3.7.1
nltk==3.9.1
numpy==2.3.2
pandas==2.3.1
protobuf==6.32.0
pydantic==2.11.7
Requests==2.32.5
scikit_learn==1.7.1
tensorflow==2.20.0
Unidecode==1.4.0
uvicorn==0.35.0